+++ atlasdb-commons/src/main/java/com/palantir/common/proxy/InterruptibleProxy.java
@@ -27,6 +27,7 @@ import com.palantir.common.base.Throwables;
 import com.palantir.common.concurrent.NamedThreadFactory;
 import com.palantir.common.concurrent.PTExecutors;
 import com.palantir.exception.PalantirInterruptedException;
+import com.palantir.remoting1.tracing.Tracers;
 
 /**
  * Proxy that calls the requested method in another thread waits on a Future.
@@ -35,20 +36,20 @@ import com.palantir.exception.PalantirInterruptedException;
  * @author dcohen
  *
  */
-public class InterruptibleProxy implements DelegatingInvocationHandler {
+public final class InterruptibleProxy implements DelegatingInvocationHandler {
 
     private final CancelDelegate cancel;
 
     @SuppressWarnings("unchecked")
     public static <T> T newProxyInstance(Class<T> interfaceClass, T delegate,
             CancelDelegate cancel) {
-        return (T)Proxy.newProxyInstance(interfaceClass.getClassLoader(),
+        return (T) Proxy.newProxyInstance(interfaceClass.getClassLoader(),
                 new Class<?>[] {interfaceClass}, new InterruptibleProxy(delegate, cancel));
     }
 
     private final Object delegate;
-    private static final ExecutorService executor = PTExecutors.newCachedThreadPool(
-            new NamedThreadFactory("Interruptible Proxy", true /* isDaemon */ ));
+    private static final ExecutorService executor = Tracers.wrap(PTExecutors.newCachedThreadPool(
+            new NamedThreadFactory("Interruptible Proxy", true /* isDaemon */ )));
 
     private InterruptibleProxy(Object delegate, CancelDelegate cancel) {
         this.delegate = delegate;